// prisma/generated/enums.ts
export const Race = {
  ASIAN: 'ASIAN',
  AFRICAN: 'AFRICAN',
  HISPANIC: 'HISPANIC',
  WHITE: 'WHITE',
  DESI: 'DESI',
  ARABIC: 'ARABIC'
} as const

export type Race = (typeof Race)[keyof typeof Race]


export const BodyType = {
  REGULAR: 'REGULAR',
  PLUS: 'PLUS',
  ATHLETE: 'ATHLETE'
} as const

export type BodyType = (typeof BodyType)[keyof typeof BodyType]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  TRANS: 'TRANS'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]